Moved the settings into src/util
