+++ src/complex/Array/SpectrumArrayItem.tsx
@@ -67,7 +67,7 @@ import ChevronUp from '@spectrum-icons/workflow/ChevronUp';
 import './SpectrumArrayItem.css';
 
 import SpectrumProvider from '../../additional/SpectrumProvider';
-import settings from '../../../settings';
+import settings from '../../util/settings';
 
 export interface OwnPropsOfSpectrumArrayItem {
   index: number;
