+++ src/complex/ArrayModal/SpectrumArrayModalItem/ModalItemHeader.tsx
@@ -16,7 +16,7 @@ import Close from '@spectrum-icons/workflow/Close';
 import ChevronUp from '@spectrum-icons/workflow/ChevronUp';
 import ChevronDown from '@spectrum-icons/workflow/ChevronDown';
 import Duplicate from '@spectrum-icons/workflow/Duplicate';
-import settings from '../../../../settings'
+import settings from '../../../util/settings';
 
 import ModalItemDelete from './ModalItemDelete';
 import FolderSearch from '@spectrum-icons/workflow/FolderSearch';
@@ -86,11 +86,13 @@ export default function ModalItemHeader({
           </ActionButton>
           <Tooltip>{childData?._path || childLabel}</Tooltip>
         </TooltipTrigger>
-
         <View>
           <Flex gap={'size-0'}>
-
-            <ActionMenu align='end' onAction={actionMenuTriggered} isQuiet={true}>
+            <ActionMenu
+              align='end'
+              onAction={actionMenuTriggered}
+              isQuiet={true}
+            >
               <Item key='delete' textValue={`delete-item-${childLabel}`}>
                 <Text>Delete</Text>
                 <Delete size='S' />
