+++ src/complex/ArrayModal/utils.tsx
@@ -39,7 +39,7 @@ import {
 import { ErrorIndicator } from '../../components/ErrorIndicator';
 import SpectrumProvider from '../../additional/SpectrumProvider';
 import { Dispatch } from './type';
-import settings from '../../../settings';
+import settings from '../../util/settings';
 
 export function getUIOptions(
   uischema: UISchemaElement,
